+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Zettelkatalog first draft.
+- Radar chart top 10 models.
+- Zettelkatalog benchmark.
 - T0066 on 2025-09-02
 - T0143 on 2025-09-02
 - T0144 on 2025-09-02
@@ -23,10 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - T0164 on 2025-09-02
 - T0165 on 2025-09-02
 - T0166 on 2025-09-02
-- 
 - T0151 on 2025-09-02
 - T0152 on 2025-09-02
 - T0155 on 2025-09-02
+
 ## [v0.2.0] - 2025-08-31
 
 ### Added